+++ test/Log.test.js
@@ -23,38 +23,73 @@ const Assert = require('../lib/Assert')
 const Log = require('../lib/Log')
 runner.suite('Log', (it) => {
   const logName = 'kado'
-  const log = new Log.LogEngine({ name: logName })
-  const logEngine = new Log(null, log)
+  const log = new Log()
+  const logEngine = new Log.LogEngine({ name: logName, toConsole: false })
+  log.addEngine('test', logEngine)
+  /* additional test code
+  const l = new Log()
+  const engine = new Log.LogEngine()
+  l.addEngine('test', engine)
+  l.info('foo')
+  l.error('bad')
+  const err = new Error('test')
+  l.dump(err)
+  l.dump(2)
+  l.dump(false)
+  l.dump({ foo: 'foo' })
+   */
   it('should construct', () => {
-    Assert.isType('Log', new Log())
+    Assert.isType('Log', log)
+  })
+  it('should dump variables', () => {
+    log.dump(true)
+  })
+  it('should dump errors', () => {
+    const error = new Error('testing')
+    log.dump(error)
+  })
+  it('should dump objects', () => {
+    const obj = { foo: 'foo' }
+    log.dump(obj)
+  })
+  it('should dump null', () => {
+    log.dump(null)
+  })
+  it('should dump numbers', () => {
+    log.dump(35)
+  })
+  it('should print', () => {
+    log.print('foo')
+  })
+  it('should log', () => {
+    log.log('foo')
   })
   it('should have no logger', () => {
-    console.log(logEngine.listEngines())
-    Assert.eq(logEngine.listEngines().length, 0)
+    Assert.eq(log.listEngines().length, 1)
   })
   it('should have no handlers', () => {
-    Assert.eq(logEngine.listEngines().length, 0)
+    Assert.eq(log.listEngines().length, 1)
   })
   it('should add an engine', () => {
-    Assert.isType('LogEngine', logEngine.addEngine(logName, log))
+    Assert.isType('LogEngine', log.addEngine(logName, logEngine))
   })
   it('should get the handler', () => {
-    Assert.isType('LogEngine', logEngine.getEngine(logName))
+    Assert.isType('LogEngine', log.getEngine(logName))
   })
   it('should activate the handler', () => {
-    Assert.isType('LogEngine', logEngine.activateEngine(logName))
+    Assert.isType('LogEngine', log.activateEngine(logName))
   })
   it('should return the active handler', () => {
-    Assert.isType('LogEngine', logEngine.getActiveEngine())
+    Assert.isType('LogEngine', log.getActiveEngine())
   })
   it('should remove the handler', () => {
-    Assert.eq(logEngine.removeEngine(logName), true)
+    Assert.eq(log.removeEngine(logName), true)
   })
   it('should no longer have the handler', () => {
-    Assert.eq(logEngine.getEngine(logName), false)
+    Assert.eq(log.getEngine(logName), false)
   })
   it('should now not have a logger', () => {
-    Assert.eq(logEngine.getActiveEngine(), null)
+    Assert.eq(log.getActiveEngine(), null)
   })
 })
 if (require.main === module) runner.execute().then(code => process.exit(code))